Genaralize dense regression net
